codex(refactor): extract compose renderer helpers
